Refactor Qwen launchers into scripts package
